pyproject.toml
[tool.poetry]
name = "lab2tfl"
version = "0.1.0"
description = ""
authors = ["RelictRLCT <pigift@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
automata-lib = "^8.4.0"
coloraide = "^4.0"
pygraphviz = "^1.14"
graphviz = "^0.20.3"
matplotlib = "^3.9.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
